requirements.txt
Django>=4.0.1,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.15.1,<0.16
Pillow>=9.1.0,<9.2
django-filter>=2.4.0,<2.5.0
celery>=4.4.7,<4.5.0
redis>=4.5.1,<4.6.0
environ==1.0